[#6] refactor: article 관련 컨트롤러와 dto 도메인 분리
