+++ src/main/java/org/sopt/controller/article/dto/SearchedArticlesResponse.java
@@ -1,4 +1,4 @@
-package org.sopt.controller.member.dto;
+package org.sopt.controller.article.dto;
 
 import java.util.List;
